moved debug code into own file
